requirements.txt
fastapi
uvicorn
python-dotenv
requests
transformers
langchain
torch
ultralytics
